Refactor App and Extension components to integrate QueryClientProvider
- Updated the App component to wrap the CoreApp with QueryClientProvider, enhancing query management.
- Refactored the Extension's index.tsx to utilize ExtensionQueryClientProvider for improved state handling.
- Introduced a new ExtensionQueryClientProvider component to manage query persistence in the extension.
- Simplified the CoreApp component by removing the queriesPersister prop, streamlining its structure.
